Change Security to OS dependent solution
The initial key storing was very insecure as it merely stored
a hidden but unencrypted key on the disk.
The new implementation uses the Data Protection (DPAPI)
from .NET - Unix and OSX are not implemented yet.
See ISecureStorage to implement other OS types.
Adjusted .csproj files to correctly display #if WINDOWS.
Ran unit tests to ensure key file is created and read correctly.